+++ app/src/ui/stashing/stash-diff-viewer.tsx
